Cast to integer

+++ transform/snowflake-dbt/models/smau_events/smau_events/manage_smau_events.sql
@@ -2,11 +2,11 @@ WITH manage_snowplow_smau_events AS (
   
   SELECT
     user_snowplow_domain_id,
-    user_custom_id      AS gitlab_user_id,
+    user_custom_id::INTEGER   AS gitlab_user_id,
     event_date,
     event_type,
-    page_view_id        AS sk_id,
-    'snowplow_frontend' AS source_type
+    page_view_id              AS sk_id,
+    'snowplow_frontend'       AS source_type
   
   FROM {{ ref('manage_snowplow_smau_events')}}
   
@@ -15,12 +15,12 @@ WITH manage_snowplow_smau_events AS (
 , manage_gitlab_dotcom_smau_events AS (
   
   SELECT
-    NULL             AS user_snowplow_domain_id,
-    user_id          AS gitlab_user_id,
+    NULL               AS user_snowplow_domain_id,
+    user_id::INTEGER   AS gitlab_user_id,
     event_date,
     event_type,
     sk_id,
-    'gitlab_backend' AS source_type    
+    'gitlab_backend'   AS source_type    
 
   FROM {{ ref('manage_gitlab_dotcom_smau_events')}}
